Refactor preferences UI and logging utilities
