+++ src/constants/times/core.ts
@@ -4,7 +4,7 @@
 
 /* Shared time constants (core) - do not import runtime gettext here. */
 
-import { GettextFunctions } from "../../utils/gettext_utils.js";
+import { GettextFunctions } from "../../utils/gettext";
 
 export function createTimeConstants(fns: GettextFunctions) {
   const { pgettext } = fns;
