+++ src/extension.ts
@@ -22,9 +22,9 @@ import { WordPack } from "./word_pack.js";
 import { Errors } from "./constants/index.js";
 import SettingsKey from "./models/settings_keys";
 import { logErr, logWarn } from "./utils/error_utils.js";
-import { fuzzinessFromEnumIndex } from "./utils/fuzziness_utils.js";
-import { createTranslatePackGetter } from "./utils/translate_pack_utils.js";
-import { extensionGettext } from "./utils/gettext_utils_ext.js";
+import { fuzzinessFromEnumIndex } from "./utils/parse";
+import { createTranslatePackGetter } from "./utils/translate";
+import { extensionGettext } from "./utils/gettext";
 import { SettingsManager } from "./services/settings_manager.js";
 import { StyleService } from "./services/style_service.js";
 import { NotificationService } from "./services/notification_service.js";
