+++ src/prefs/formatters.ts
@@ -1,0 +1,64 @@
+import Gtk from "gi://Gtk";
+import { ClockFormatter, TimeFormat } from "../core/clock_formatter.js";
+import { fuzzinessFromEnumIndex } from "../utils/parse/index.js";
+import SettingsKey from "../models/settings_keys.js";
+import TRANSLATE_PACK from "./translate_pack.js";
+
+export function getTimeFormatsList(settings: any): Gtk.StringList {
+  /**
+   * Build a localized list of example clock strings for the time format chooser.
+   *
+   * This uses the clock formatter and current fuzziness setting to render two
+   * example time strings which are displayed in the preferences combo row.
+   *
+   * @param settings - Gio.Settings to read current fuzziness setting
+   * @returns Gtk.StringList - list containing two localized example strings
+   */
+  const clockFormatter = new ClockFormatter(TRANSLATE_PACK());
+  const date = new Date();
+  const fuzzinessEnumIndex = settings.get_enum(SettingsKey.FUZZINESS);
+  const fuzziness = fuzzinessFromEnumIndex(fuzzinessEnumIndex);
+
+  const timeFormatOne = clockFormatter.getClockText(
+    date,
+    false,
+    false,
+    TimeFormat.FORMAT_ONE,
+    fuzziness,
+  );
+  const timeFormatTwo = clockFormatter.getClockText(
+    date,
+    false,
+    false,
+    TimeFormat.FORMAT_TWO,
+    fuzziness,
+  );
+
+  return new Gtk.StringList({ strings: [timeFormatOne, timeFormatTwo] });
+}
+
+export function addTimeFormatComboRow(group: any, settings: any) {
+  /**
+   * Create and add a time format combo row to a group.
+   *
+   * The combo shows two localized example strings generated by
+   * `getTimeFormatsList` and binds selection back to the TIME_FORMAT setting.
+   *
+   * @param group - Preferences group to add the combo to
+   * @param settings - Gio.Settings used to read/write the TIME_FORMAT value
+   * @returns Adw.ComboRow - the created combo row
+   */
+  const info = {
+    title: "Time Format",
+    subtitle: "Choose time format",
+    model: getTimeFormatsList(settings),
+    selected: settings!.get_enum(SettingsKey.TIME_FORMAT),
+  };
+  // Use Adw.ComboRow via dynamic import at runtime in prefs wrapper
+  const row = new (globalThis as any).Adw.ComboRow(info);
+  group.add(row);
+  row.connect("notify::selected", (widget: any) =>
+    settings!.set_enum("time-format", widget.selected),
+  );
+  return row;
+}
