+++ src/tests/setup.ts
@@ -8,7 +8,7 @@
  */
 
 import { setLogger } from "../utils/error_utils.js";
-import { consoleLogger } from "../utils/logger_console.js";
+import { consoleLogger } from "../utils/logging";
 
 // Add String.prototype.format to match GNOME Shell environment
 declare global {
