+++ src/tests/unit/constants/times/core.test.ts
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 import { createTimeConstants } from "../../../../constants/times/core.js";
-import type { GettextFunctions } from "../../../../utils/gettext_utils.js";
+import type { GettextFunctions } from "../../../../utils/gettext";
 
 describe("createTimeConstants", () => {
   let mockGettextFunctions: GettextFunctions;
