+++ src/utils/color/color_utils.ts
@@ -1,23 +1,7 @@
-/*
- * SPDX-FileCopyrightText: 2024 Wesley Benica <wesley@benica.dev>
- * SPDX-License-Identifier: GPL-3.0-or-later
- */
-
-/**
- * Utility functions for color handling
- */
-
-/**
- * Normalizes a color string to hex format
- *
- * @param color - The color string to normalize
- * @returns The normalized hex color string
- */
+/* Utility functions for color handling (moved) */
 export function normalizeColor(color: string): string {
   if (!color) return "#ffffff";
   color = color.trim();
-
-  // Handle rgb() format (allow negative numbers, clamp to [0,255])
   const rgbMatch = color.match(
     /rgb\s*\(\s*(-?\d+)\s*,\s*(-?\d+)\s*,\s*(-?\d+)\s*\)/i,
   );
@@ -27,22 +11,11 @@ export function normalizeColor(color: string): string {
     const b = Math.max(0, Math.min(255, Number(rgbMatch[3])));
     return "#" + [r, g, b].map((n) => n.toString(16).padStart(2, "0")).join("");
   }
-
-  // Handle hex format (with or without #)
   const hexMatch = color.match(/^#?[0-9a-f]{3,6}$/i);
-  if (hexMatch) {
-    return color.startsWith("#") ? color : `#${color}`;
-  }
-
-  return "#ffffff"; // fallback
+  if (hexMatch) return color.startsWith("#") ? color : `#${color}`;
+  return "#ffffff";
 }
 
-/**
- * Escapes markup characters for Pango markup
- *
- * @param text - The text to escape
- * @returns The escaped text
- */
 export function escapeMarkup(text: string): string {
   return text
     .replace(/&/g, "&amp;")
