+++ src/utils/gettext/gettext_utils.ts
