+++ src/utils/logging/logger_console.ts
@@ -3,16 +3,17 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-/**
- * Console/Node.js logger implementation for tests
- */
 import type { Logger } from "./logger_interface.js";
 
 export const consoleLogger: Logger = {
   log(...args: unknown[]): void {
-    console.log(...args);
+    if (typeof (globalThis as any).console !== "undefined") {
+      (globalThis as any).console.log(...args);
+    }
   },
   logError(error: Error, message?: string): void {
-    console.error(message || error.message, error);
+    if (typeof (globalThis as any).console !== "undefined") {
+      (globalThis as any).console.error(message || error.message, error);
+    }
   },
 };
