+++ src/utils/parse/fuzziness_utils.ts
@@ -1,12 +1,6 @@
-import { Fuzziness } from "../core/clock_formatter.js";
-import { FUZZINESS_ENUM_MINUTES } from "../constants/index.js";
+import { Fuzziness } from "../../core/clock_formatter.js";
+import { FUZZINESS_ENUM_MINUTES } from "../../constants/index.js";
 
-/**
- * Maps a fuzziness enum index to its corresponding Fuzziness value
- *
- * @param index - The index from the settings enum
- * @returns The corresponding Fuzziness value
- */
 export function fuzzinessFromEnumIndex(index: number): Fuzziness {
   const minutes = FUZZINESS_ENUM_MINUTES[index] ?? 5;
   switch (minutes) {
@@ -23,12 +17,6 @@ export function fuzzinessFromEnumIndex(index: number): Fuzziness {
   }
 }
 
-/**
- * Parses a fuzziness value from string or number to Fuzziness enum
- *
- * @param value - The value to parse
- * @returns The parsed Fuzziness value
- */
 export function parseFuzziness(value: Fuzziness | string): Fuzziness {
   if (typeof value === "string") {
     const parsed = parseInt(value);
@@ -38,7 +26,7 @@ export function parseFuzziness(value: Fuzziness | string): Fuzziness {
     ) {
       return parsed as Fuzziness;
     }
-    return Fuzziness.FIVE_MINUTES; // fallback
+    return Fuzziness.FIVE_MINUTES;
   }
   return value;
 }
